Debugging

+++ test.sh
@@ -1,8 +1,8 @@
 for i in "$@"
 do
 case $i in
-    -r=*|--restart=*)
-    RESTART="${i#*=}"
+    -r|--restart)
+    RESTART="1"
     ;;
 
     -s=*|--searchpath=*)
@@ -18,30 +18,29 @@ case $i in
     ;;
 
     *)
-            # unknown option
     ;;
 esac
 done
-echo RESTART = ${RESTART}
 
 
 declare -a scripts=("420bits-service.py" "420bits-webserver.py")
 for script in "${scripts[@]}"
 do
    echo "$script"
-   
-	PID=`ps -eaf | grep $script | grep -v grep | awk '{print $2}'`
-	if [[ "" !=  "$PID" ]]; then
-	  echo "killing $PID"
-	  kill -9 $PID
-	fi
+   	if [ "$RESTART" -eq "1" ]; then
+   		PID=`ps -eaf | grep $script | grep -v grep | awk '{print $2}'`
+		if [[ "" !=  "$PID" ]]; then
+		  echo "killing $PID"
+		  kill -9 $PID
+		fi
+   	fi
 
 	if (( $(ps -aux | grep -v grep | grep "$script" | wc -l) > 0 ))
 	then
-	echo "420bits-service.py is already running"
+	echo "$script is already running"
 	else
-	echo "Will start 420bits-service.py"
+	echo "Will start $script"
 	/usr/bin/python "/home/pi/420bits/local-sever/$script" > "/home/pi/420bits/logs/$script.log" 2>&1 &
-	echo "Did start 420bits-service.py"
+	echo "Did start $script --"
 	fi
 done
